debt - rename `titleControl` to `editorTabsControl`
