+++ src/vs/workbench/browser/parts/editor/media/editortabscontrol.css
